Drizzle.jl

Drizzle.jl implements the drizzle (variable-pixel linear reconstruction) algorithm used to stack image sequences, most commonly used in astronomical imaging.

Drizzle.AbstractDroplet — Type

AbstractDroplet{T}

Supertype for all droplets (drizzle kernels).

All droplets should define Drizzle.pixfrac(::K), which describes the linear scale of the droplet relative to that of the pixels in the input image (i.e. pixel fraction). pixfrac(::AbstractDroplet) defaults to k.pixfrac.

Drizzle.GaussianDroplet — Type

GaussianDroplet{T<:Real} <: AbstractDroplet{T}

A droplet based on a Gaussian function with pixfrac equal to the ratio of the full-width half maximum (FWHM) to the pixel length scale.

Although this droplet is not flux-preserving like SquareDroplet, it tends to preserve point spread functions better and may yield a smoother result, which may be more appealing for an astrophotographer, or even point source photometry.

Drizzle.LanczosDroplet — Type

LanczosDroplet{A,T<:Real} <: AbstractDroplet{T}

A droplet based on the Lanczos window function.

!!! note This droplet is only designed for resampling images without rescaling them (i.e. rotating and reflecting). pixfrac should usually be 1.

Drizzle.PointDroplet — Type

PointDroplet <: AbstractDroplet{Bool}

The simplest droplet: an infinitely small point. This is equivalent to any other droplet k with pixfrac(k) == 0.

In practice, this droplet requires large numbers of dither positions to avoid leaving gaps in the image, but it is fast.

Drizzle.SquareDroplet — Type

SquareDroplet{T<:Real} <: AbstractDroplet{T}

A droplet which maps the input pixels to centered squares with the same orientation as the input. This droplet is flux-preserving and suitable for stellar photometric applications, but may not preserve PSFs as well as GaussianDroplet.

For faster processing, consider using TurboDroplet instead, or PointDroplet if there are a large number of dither positions and pixfrac is very small.

Drizzle.TurboDroplet — Type

TurboDroplet{T<:Real} <: AbstractDroplet{T}

A simplified implementation of SquareDroplet. This droplet assumes that the rotation of the input image is negligible compared to the output grid, which can speed up processing for large datasets.

Drizzle._pixfrac_check — Method

_pixfrac_check(pixfrac)

Checks that the pixel fraction is not negative.

Drizzle.drizzle! — Method

drizzle!(transform, grid!, image, droplet::AbstractDroplet)

Drizzles a image onto grid! with the given droplet model. This function may be called with do syntax describing the transform function.

Drizzle.is_flux_preserving — Method

is_flux_preserving(::Type{<:AbstractDroplet})
is_flux_preserving(::AbstractDroplet)

Returns true if the kernel is flux-preserving, false otherwise. For new AbstractDroplet subtypes, this defaults to false.

Drizzle.pixfrac — Method

pixfrac(k::AbstractDroplet)

Gets the linear scale factor (pixel fraction) for the given droplet.

Note that pixfrac is not necessarily comparable across different types of droplets. For droplets with a finite support (like SquareDroplet), pixfrac is the linear scaling factor relative to the support's bounding box. Droplets without finite support (like GaussianDroplet) should have their pixfrac convention documented.